Update install info for Steam users
The linked 2K Support article about reverting to the GameSpy version of BtS stopped working a few months ago, apparently due to a login wall. Now linking to an archived version of the article. And it's not clear that the default Steam version really doesn't work with Taurus, so I've rephrased the install instructions accordingly. (It is known that the runtime code modifications won't work with the default Steam version, i.e. part of the savegame import/export functionality and the balloon size options.)